Re structured project to allow maintainability and easy code refactoring. Create class library for Member, DataBase and LogManagement

+++ CreditUnionFYP/classes/User.cs
@@ -8,7 +8,7 @@ namespace CreditUnionFYP
 {
     public class User
     {
-        private int userId;
+        private byte userId;
         public string userName;
         private string salt;
         private string pwd;
